+++ src/Router.php
@@ -188,7 +188,7 @@ class Router
         $trace = debug_backtrace();
         $class = get_class($this);
         trigger_error("Call to undefined method $class::$method from {$trace[0]['file']} " .
-                      "on line {$trace[0]['line']}", E_USER_ERROR);
+                      "on line {$trace[0]['line']}", E_USER_WARNING);
     }
 
     /**